[WIP] Resolve compile errors

+++ src/block.rs
@@ -12,10 +12,9 @@ use chrono::prelude::*;
 
 #[derive(Serialize, Deserialize, Debug,Clone)]
 pub struct Header {
-    //pub parenthash: H256,
     pub nonce: u32,
     pub difficulty: H256,
-    pub timestamp: i32,
+    pub timestamp: u128,
     pub merkle_root:H256,
     pub miner_id:i32,
 }
@@ -46,7 +45,7 @@ pub struct Block {
 impl Block {
     pub fn new(
         //parent: H256,
-        ts: i32,
+        ts: u128,
         n: u32,
         content_merkle_root: H256,
         sortition_proof: Vec<H256>,
